requirements.txt
fastapi
uvicorn[standard]
transformers
torch
peft
bitsandbytes
pandas
openai-whisper